DROID-3628 refactoring
