+++ app/src/main/java/com/anytypeio/anytype/di/feature/notifications/PushDI.kt
@@ -1,11 +1,21 @@
 package com.anytypeio.anytype.di.feature.notifications
 
+import android.app.NotificationManager
+import android.content.Context
 import com.anytypeio.anytype.device.AnytypePushService
+import com.anytypeio.anytype.device.DefaultPushMessageProcessor
+import com.anytypeio.anytype.device.NotificationBuilder
+import com.anytypeio.anytype.device.PushMessageProcessor
 import com.anytypeio.anytype.di.common.ComponentDependencies
 import com.anytypeio.anytype.domain.device.DeviceTokenStoringService
+import com.anytypeio.anytype.presentation.notifications.CryptoService
+import com.anytypeio.anytype.presentation.notifications.CryptoServiceImpl
 import com.anytypeio.anytype.presentation.notifications.DecryptionPushContentService
+import com.anytypeio.anytype.presentation.notifications.DecryptionPushContentServiceImpl
+import com.anytypeio.anytype.presentation.notifications.PushKeyProvider
 import dagger.Component
 import dagger.Module
+import dagger.Provides
 import javax.inject.Singleton
 
 @Singleton
@@ -24,10 +34,56 @@ interface PushContentComponent {
 
 @Module
 object PushContentModule {
+    @JvmStatic
+    @Provides
+    @Singleton
+    fun provideNotificationManager(
+        context: Context
+    ): NotificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
 
+    @JvmStatic
+    @Provides
+    @Singleton
+    fun provideNotificationBuilder(
+        context: Context,
+        notificationManager: NotificationManager
+    ): NotificationBuilder = NotificationBuilder(
+        context = context,
+        notificationManager = notificationManager
+    ).apply {
+        createNotificationChannelIfNeeded()
+    }
+
+    @JvmStatic
+    @Provides
+    @Singleton
+    fun providePushMessageProcessor(
+        decryptionService: DecryptionPushContentService,
+        notificationBuilder: NotificationBuilder
+    ): PushMessageProcessor = DefaultPushMessageProcessor(
+        decryptionService = decryptionService,
+        notificationBuilder = notificationBuilder
+    )
+
+    @JvmStatic
+    @Provides
+    @Singleton
+    fun provideCryptoService(): CryptoService = CryptoServiceImpl()
+
+    @JvmStatic
+    @Provides
+    @Singleton
+    fun provideDecryptionPushContentService(
+        pushKeyProvider: PushKeyProvider,
+        cryptoService: CryptoService,
+    ): DecryptionPushContentService = DecryptionPushContentServiceImpl(
+        pushKeyProvider = pushKeyProvider,
+        cryptoService = cryptoService,
+    )
 }
 
 interface PushContentDependencies : ComponentDependencies {
     fun deviceTokenSavingService(): DeviceTokenStoringService
-    fun decryptionService(): DecryptionPushContentService
+    fun pushKeyProvider(): PushKeyProvider
+    fun context(): Context
 }
